requirements.txt
mcp
openai
python-dotenv
pytest
pytest-asyncio